requirements.txt
fastapi
uvicorn
langgraph
langchain-community
duckduckgo-search
ddgs
google-auth
google-genai
